pyproject.toml
[project]
name = "master_thesis"
version = "0.0.1"
description = "A short description of the project."
authors = [{ name = "Jonas Hoffmann" }, { name = "Joachim Schrøder Andersson" }]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "datasets==3.6.0",
  "huggingface-hub>=0.34.4",
  "ipykernel>=6.30.1",
  "jupyter>=1.1.1",
  "kenlm>=0.3.0",
  "librosa>=0.11.0",
  "loguru>=0.7.3",
  "numpy>=1.26.4",
  "pip>=25.2",
  "pyctcdecode>=0.5.0",
  "python-dotenv>=1.1.1",
  "ruff>=0.12.11",
  "rust-just>=1.43.0",
  "soundfile>=0.13.1",
  "torch>=2.8.0",
  "torchcodec>=0.7.0",
  "tqdm>=4.67.1",
  "transformers>=4.56.2",
  "typer>=0.19.2",
  "evaluate>=0.4.2",
  "jiwer>=4.0.0",
  "hydra-core>=1.3.2",
  "pydantic>=2.11.10",
  "subjob",
  "wandb>=0.22.1",
]
requires-python = "~=3.10.0"

[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"] # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[tool.uv.sources]
subjob = { git = "ssh://git@github.com/jhoffe/subjob" }

[tool.uv]
package = true
